load game now creates new game if save data doesn't exist

+++ src/main/java/connectFour/ConnectFourControllerView.java
@@ -200,17 +200,22 @@ public class ConnectFourControllerView extends GameControllerView {
 	 * @return true if the load was successful, false otherwise
 	 */
 	public boolean loadSaveGame() {
+		boolean retVal = true;
 		try {
 			String fname = accountmanager.getCurUsername() + "-" + gameName + ".dat";
 			String sep = System.getProperty("file.separator");
 			String filepath = System.getProperty("user.dir") + sep + "save-data" + sep + fname;
 			File file = new File(filepath);
-			FileInputStream fis = new FileInputStream(file);
-			ObjectInputStream ois = new ObjectInputStream(fis);
-			gameModel = (ConnectFourModel) ois.readObject();
-			ois.close();
-			update(gameModel, this);
-			file.delete();
+			if(file.exists()) {
+				FileInputStream fis = new FileInputStream(file);
+				ObjectInputStream ois = new ObjectInputStream(fis);
+				gameModel = (ConnectFourModel) ois.readObject();
+				ois.close();
+				update(gameModel, this);
+				file.delete();
+			} else {
+				retVal = newGame();
+			}
 		} catch(IOException | ClassNotFoundException e) {
 			return false;
 		}
@@ -218,7 +223,7 @@ public class ConnectFourControllerView extends GameControllerView {
 		setupListeners();
 		gameModel.addObserver(this);
 		update(gameModel, this);
-		return true;
+		return retVal;
 	}
 
 	@Override
